+++ NEWS.md
@@ -3,7 +3,7 @@
 ## Issues
 
 - Internal function `is_doc_open()` replaces `is_office_doc_edited()` to check if a document is open on Windows. This solves the issue where RStudio crashed while trying to write to an open Office document.
-
+- Add option `preserve` to preserve cell line breaks to `docx_summary()`.
 
 # officer 0.6.2